Decouple quality components from sample views
Allows reuse in cache views.
