+++ client/src/js/quality/components/Chart.js
